:new: :memo: documentation: Add new page, "Frontier technology tips"
This commit adds a new page with tips and advice about writing
documentation for emerging / frontier technology. Since this is
something very familiar to UNICEF Innovation Fund projects, it makes
sense to include this in the Heuristics site.

It is adapted from a summary in a blog post I published in February:

https://blog.jwf.io/2020/02/devconf-cz-2020-play-by-play/

Signed-off-by: Justin W. Flory <git@jwf.io>